feat(app/worker-profile/my-jobs): fix according to review
